Support multiarch builds to run on other architectures e.g. Raspberry Pi
Note: This depends on the image nimmis/ubuntu being rebuilt with multiarch support

This uses docker buildx and manifest lists to build a multiple versions of an image
for different systems which appear as a single image on the docker hub.
This allows clients to select the right version automatically.

Changes:
- Select the right cpu arch version when downloading java
- Replace the java install in the Dockerfile with a call to set_java_ver
- README updated on how to build these images